move functionality over to footer
